+++ src/main/java/net/thedevilthing/firstmod/datagen/ModItemTagProvider.java
@@ -1,6 +1,7 @@
 package net.thedevilthing.firstmod.datagen;
 
 import net.minecraft.core.HolderLookup;
+import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.data.PackOutput;
 import net.minecraft.data.tags.ItemTagsProvider;
 import net.minecraft.resources.ResourceLocation;
@@ -9,10 +10,13 @@ import net.minecraft.world.item.Items;
 import net.minecraft.world.level.block.Block;
 import net.neoforged.neoforge.common.data.ExistingFileHelper;
 import net.thedevilthing.firstmod.Firstmod;
+import net.thedevilthing.firstmod.InfusionItemsConfig;
 import net.thedevilthing.firstmod.util.ModTags;
 import org.jetbrains.annotations.Nullable;
 
+import java.util.Set;
 import java.util.concurrent.CompletableFuture;
+import java.util.stream.Collectors;
 
 public class ModItemTagProvider extends ItemTagsProvider {
     public ModItemTagProvider(PackOutput output, CompletableFuture<HolderLookup.Provider> lookupProvider, CompletableFuture<TagLookup<Block>> blockTags, @Nullable ExistingFileHelper existingFileHelper) {
@@ -21,10 +25,5 @@ public class ModItemTagProvider extends ItemTagsProvider {
 
     @Override
     protected void addTags(HolderLookup.Provider provider) {
-        tag(ModTags.Items.INFUSABLES)
-                .add(Items.REDSTONE)
-                .addOptionalTag(ResourceLocation.fromNamespaceAndPath("c", "ingots"))
-                .addOptionalTag(ResourceLocation.fromNamespaceAndPath("c", "gems"))
-                .remove(ResourceLocation.fromNamespaceAndPath("minecraft", "prismarine_crystals"));
     }
 }